auth-backend: move prepareBackstageIdentityResponse to auth-node
